BASIC CATEGORY AND TEST ROUTES ARE NOW WORKING
The many-to-many relationship is now working between categories and
tests. The tests don't yet have questions and choices. That will be a
future commit. This commit focused on establishing the m2m
relationship.

I created the “seed_data” folder with the data.sql file which loads
some sample data so that I don’t need to use the admin screen.

I added the Category and Test models to the admin.py file so that the
admin screen may be used if necessary to create categories and tests
versus writing SQL statements.

I added the Category and Test models. In models.py. The many-to-many
relationship is working. Both models have unique constraints on them.
The test model does not yet have the questions and choices tables.
I need to override the User model in a future commit.

In the api/urls.py I have the initial routes working but they may
change.

In views.py there current exists only simple Category and Test view
sets. These may change in the future.

Finally, the first migration file was created with basic test and
category routes working.