A TEST CAN NOW BE CREATED SANS OPTIONS
In serializers.py, I overrode the create method for the TestSerializer
class. This was done since this commit now can save a test with a
collection of questions. And because of the many-to-many relationship
between Tests and Categories, the create method needs to be
overwritten.

The questions do not yet have the choices. That will be a future
commit. This commit is focused on the POSTing of a test with questions.

I added support for transactions from “djabngo.db”. That is because
posting a test needs to be atomic.

The bulk of what has changed is in the create method for the test
serializer.